feat: add entities

+++ SumduDataVaultApi/SumduDataVaultApi/DataAccess/Entities/User.cs
@@ -2,7 +2,7 @@
 
 namespace SumduDataVaultApi.DataAccess.Entities
 {
-    public class User : IdentityUser<int>
+    public class User : IdentityUser<long>
     {
         public required string FirstName { get; set; }
         public required string LastName { get; set; }
